Refactor application section
